+++ Clinic.Reports/DoctorActivityReport.cs
@@ -11,7 +11,7 @@ namespace Clinic.Reports
         public DateTime StartDateTime { get; private set; }
         public DateTime EndDateTime { get; private set; }
         public int ExaminationsCount { get; private set; }
-        public double TotalAmountReceived { get; private set; }
+        public decimal TotalAmountReceived { get; private set; }
         
         private DoctorActivityReport() { }
